Updating README to work with the rust example
Hello! 
I followed the example for rust and noticed I got this error in the browser: 
`:8000/site/wasm_demo.wasm:1 Failed to load resource: the server responded with a status of 404 (File not found)`
And it looks like site.js is referencing wasm_demo.wasm which got renamed to site.wasm in the Makefile example. 
Once I updated the Makefile to keep the name wasm_demo and reference that instead, the example worked.